Major rewrite of logging and config.
Logging facilities (getLog() method and such) have been moved to
logClient.py to better reflect their dependency on the new
logServer.py. logServer.py implements a logger, which listens (in a
parallel thread) to SocketHandler connections from another logger and
writes a log file to disk (currently in 'output.log'). This has been
done so that other processes, started by neuroConstruct (for example the
MultiConductance.py) can log to the same file without interfering with
one another. it is partially configurable via the general.cfg
file. Logging using the logClient.py is send to the console screen and
the logServer, if running.

The parameters for configuring the multiEA.py run have been all made
available in the general.cfg and they should only be edited in this file
now.
This allows for easy manipulation using another config parser (in Qt
for example).