Added out of folder simulation and result storing.
Now, by default, the simulations will be created in a new temp directory
for a cleaner programme folder. This is especially important, since the
ROB server has a really slow home, but fast temp directory.
    In addition, the results will be copied to an extra folder after
execution.

The configuration file must be specified when running multiEA.py via the
command switch "-c <Config>". To make managing several configs easier.

The project configuration will still be read with projConf.py, but is no
longer global (for the module). That means that the ProjConf instance
must be passed to other methods or objects if they rely heavily on
project configurations (preferably via arguments in constructors).