Made the "general.cfg" to "default.cfg".
That means that the new "default.cfg" will ALWAYS be parsed before the
user specified config (via -c or --config). This will make changing
settings easier and does not break so easily on switching revisions of
"default.cfg".
So please don't push any custom versions of the default configuration to
the repo if it should not really affect the default behaviour for all
users. In fact don't edit it at all except for changing the default
behaviour of this program.