The simulation path is now under results.
Until now the default was to create a temp folder, but this was not the
best solution because of data loss or log files missing the last lines
because of the copying. It was also possible to force "in source" builds
that messed up the neuroConstruct project templates.

So finally the simulations have been moved for more cleanliness
and are by default under the results subdirectory (configurable).

The previous default can still be achieved using the -t or --temp option
to create the tmp folder before moving the results to the results
subdirectory.