Restructured default.cfg and exposed inspyred.
The configuration now allows for use of custom and predefined callables,
which are sections specifying the function (or callable object creation)
and the parameters that should be passed to insypred.

The logServer.py now supports automatic port acquisition, BUT only when
the log_server_port setting is 0 in both, default.cfg and the
customizations.

Also renamed some variables and cleaned main_program.py.